+++ MoonShell/Program.cs
@@ -15,7 +15,7 @@ namespace MoonShell
 
         // Enter current version here
         internal readonly static float Major = 2;
-        internal readonly static float Minor = 0;
+        internal readonly static float Minor = 1;
 
         /* END OF VERSION PROPERTIES */