+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.1] - 2017-12-22
+- Save SSH configuration upon exiting (except password)
+- Paste only text, without formatting fixed
+
 ## [2.0] - 2017-12-20
 - Added SSH support